Worldpay: Adjust use of normalized stored credentials hash
Previously the new/beta normalized stored credential hash was using a
boolean value to indicate if a transaction was part of a recurring
transaction or not. However, this did not allow for the case of gateways
(like Worldpay) to take into account transactions that are part of an
installment. This adjusts the previously "recurring" field to use the
name "transaction_type" which accepts a string of either:

- "installment"
- "recurring"
- "unscheduled"

This should allow more flexibility to for other gateways that also have
the notion of a payment that is part of an installment.

Note that this normalized hash is still in "beta" use. The hash
structure itself may change and until there is solid use around it
across multiple gateways, it will likely be a defined-yet-not-enforced
hash rather than an object which provides guarantees around field
values.

Closes #3139