Adds ELO card type
This adds ELO bin ranges and ELO brand logic in general. ELO is in a
partnership with Discover and is part of the Discover Global Network.
Using the ELO BIN ranges before the Discover regex allows those cards
to be identified correctly.

It also adds `elo` as a supported cardtype for the Adyen gateway.

Unit Tests
27 tests, 132 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications 100% passed

Remote Tests
45 tests, 122 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications 100% passed