WorldPay: Fix element order for 3DS + stored cred
WorldPay's XML DTD requires that the `session` comes after
`storedCredentials`. When using stored credential options with a 3DS
request we were ending up with the elements being generated in the wrong
order. This fixes the issue and ensures `storedCredentials` precede
`session` when both are to be included in the request document.

ECS-219

Unit:
44 tests, 242 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote
35 tests, 154 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Closes #3172