Adyen: Support adjust action
This adds support for Adyen's adjustAuthorization action, which can
change the amount of a prior authorization, either increasing or
decreasing the amount and extending the auth period. There is no other
precedent for this type of action in ActiveMerchant, but it seems very
straightforward and simple to support, similar to other followup actions
like refunds and voids.

Closes #3190

Remote:
55 tests, 163 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
34 tests, 161 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed