Do not consider a refund unsuccessful if only refunding the fee failed (#3188)
Fixes https://github.com/activemerchant/active_merchant/issues/3116

When specifying the `refund_fee_amount` on a Refund using the Stripe
gateway, Active Merchant makes three different API requests:

1. Issue the actual refund against the charge
2. Fetch the charge to get the application fee identifier
3. Issue the application fee refund against the original application fee

Currently, if either 2 or 3 fail while 1 succeeds, the `#refund`
method's response is a failure, which is not true. This process is not
transactional. The refund had succeeded and is not rolled back, which
isn't possible to do anyway.