TransFirst Express: Fix blank address2 values
If a user passes in an empty string for the address2 field, their 
transaction will fail on this gateway. This PR ensures that address2 
only gets passed through if it is present and not blank. Additionally 
it cleans up some remote test failures. test_successful_verify still 
fails in the remote tests, but failure is unrelated to change.

Unit:
21 tests, 103 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
33 tests, 107 assertions, 1 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
96.9697% passed

Closes #3231