WorldPay: Add support for store method
We are only supporting shopper-scoped tokens right now, and not
implementing delete/unstore, saving a token along with an authorize or
purchase call, overriding token details on use, or setting the token
lifetime.

The design of the opaque token authorization string we return from
`store` has been intentionally designed to be flexible enough to extend
in a backwards-compatible way should any of the above features need to
be implemented.

Unit:
66 tests, 392 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
50 tests, 219 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
96% passed

The two failing remote tests are concerning 3DS parameter pass through
which the test credentials I'm using does not have permission to use
presently.

ECS-347

Closes #3232